Implemented patients_management module
Added patients_management module to handle roles of patients and doctors in the system. This includes the creation of models, signals for group assignments, corresponding views, and templates. Also, the app was added in settings.py and urls were included in the main routing file. Further, the address autocomplete feature is integrated for registering patients.